+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * [Issue #87](https://github.com/manheim/terraform-pipeline/issues/87) Add travis CI support to the project
+* [Issue #186](https://github.com/manheim/terraform-pipeline/issues/186) Add convenience method to detect the version of terraform
 
 # v5.3